Section 15 - Deployment/requirements.txt
# *NOTES*
# These are the dependencies for the Django application, including those needed for AWS deployment with boto3.
# This file lists the packages and their versions required to run the Django project.
# It can be generated by stopping the Django server and running the command:
# pip freeze > requirements.txt
# https://pip.pypa.io/en/stable/reference/pip_freeze/


asgiref==3.3.1
# *NOTES*
# `asgiref` is a package that provides ASGI (Asynchronous Server Gateway Interface) utilities.
# ASGI is a standard for Python asynchronous web apps and servers to communicate.
# https://pypi.org/project/asgiref/

boto3==1.16.61
# *NOTES*
# `boto3` is the Amazon Web Services (AWS) SDK for Python, which allows Python developers to write software that makes use of Amazon services.
# https://pypi.org/project/boto3/

botocore==1.19.61
# *NOTES*
# `botocore` is a low-level, data-driven core of boto3. It provides the service definitions used by boto3.
# https://pypi.org/project/botocore/

Django==3.1.5
# *NOTES*
# `Django` is a high-level Python web framework that encourages rapid development and clean, pragmatic design.
# Version 3.1.5 of Django is specified.
# https://pypi.org/project/Django/3.1.5/

django-storages==1.11.1
# *NOTES*
# `django-storages` is a collection of custom storage backends for Django, including support for Amazon S3.
# https://pypi.org/project/django-storages/

jmespath==0.10.0
# *NOTES*
# `jmespath` is a query language for JSON. It allows for declaratively extracting data from JSON documents.
# https://pypi.org/project/jmespath/

Pillow==8.1.0
# *NOTES*
# `Pillow` is a fork of the Python Imaging Library (PIL) that adds some user-friendly features.
# https://pypi.org/project/Pillow/

psycopg2-binary==2.8.6
# *NOTES*
# `psycopg2-binary` is a PostgreSQL adapter for Python. It is the most popular PostgreSQL adapter for the Python programming language.
# https://pypi.org/project/psycopg2-binary/

python-dateutil==2.8.1
# *NOTES*
# `python-dateutil` provides powerful extensions to the standard datetime module, including parsing of dates in almost any string format.
# https://pypi.org/project/python-dateutil/

pytz==2020.5
# *NOTES*
# `pytz` brings the Olson tz database into Python and allows accurate and cross-platform timezone calculations.
# https://pypi.org/project/pytz/

s3transfer==0.3.4
# *NOTES*
# `s3transfer` is a library for managing Amazon S3 transfers. It is used internally by `boto3`.
# https://pypi.org/project/s3transfer/

six==1.15.0
# *NOTES*
# `six` is a Python 2 and 3 compatibility library. It provides utility functions for writing code that is compatible with both Python versions.
# https://pypi.org/project/six/

sqlparse==0.4.1
# *NOTES*
# `sqlparse` is a non-validating SQL parser for Python. It provides support for parsing, splitting, and formatting SQL statements.
# https://pypi.org/project/sqlparse/

urllib3==1.26.3
# *NOTES*
# `urllib3` is a powerful, user-friendly HTTP client for Python. It provides support for HTTP connections with thread-safe connection pooling, file post, and more.
# https://pypi.org/project/urllib3/